+++ dl-api-csharp/Doublelives.Domain/Sys/SysDict.cs
@@ -8,7 +8,7 @@ namespace Doublelives.Domain.Sys
 
         public string Num { get; set; }
 
-        public int? Pid { get; set; }
+        public long? Pid { get; set; }
 
         public string Tips { get; set; }
     }
