+++ dl-api-csharp/Doublelives.Domain/Sys/SysLoginLog.cs
@@ -20,6 +20,6 @@ namespace Doublelives.Domain.Sys
 
         public string Succeed { get; set; }
 
-        public int? Userid { get; set; }
+        public long? Userid { get; set; }
     }
 }
