+++ dl-api-csharp/Doublelives.Domain/Sys/SysRole.cs
@@ -4,16 +4,16 @@ namespace Doublelives.Domain.Sys
 {
     public class SysRole : AuditableEntityBase
     {
-        public int? Deptid { get; set; }
+        public long? Deptid { get; set; }
 
         public string Name { get; set; }
 
-        public int? Num { get; set; }
+        public long? Num { get; set; }
 
-        public int? Pid { get; set; }
+        public long? Pid { get; set; }
 
         public string Tips { get; set; }
 
-        public int? Version { get; set; }
+        public long? Version { get; set; }
     }
 }
